Refactor BLEService to WaterLevelBLE for clarity and consistency
- Renamed BLEService class to WaterLevelBLE across all relevant files.
- Updated constructor and method signatures to reflect the new class name.
- Adjusted server and pump callback classes to use WaterLevelBLE.
- Ensured all instances of BLEService are replaced in main application logic.

+++ src/wifi_manager.cpp
@@ -90,7 +90,7 @@ bool WiFiManager::reconnect() {
     
     // Exponential backoff for reconnect interval
     if (reconnectAttempts > 3) {
-        reconnectInterval = min(reconnectInterval * 2, 300000UL); // Max 5 minutes
+        reconnectInterval = min((uint32_t)(reconnectInterval * 2), (uint32_t)300000UL); // Max 5 minutes
     }
     
     return connect();
